+++ public/index.html
@@ -3,7 +3,13 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="icon" type="image/png" href="./Link.png">
+    <link rel="icon" type="image/png" href="./Link.png" />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700&family=Oswald:wght@700&family=Passion+One:wght@700&display=swap"
+      rel="stylesheet"
+    />
     <title>Linkr</title>
   </head>
   <body>
