+++ src/components/App.js
@@ -1,24 +1,29 @@
-import { BrowserRouter, Switch, Route } from 'react-router-dom'
-import React from "react";
-import "../reset-css/reset.css"
+import { BrowserRouter, Switch, Route } from "react-router-dom";
+import React, { useContext, useState } from "react";
 import PageLogin from "../Pages/PageLogin";
-import PageRegistration from '../Pages/PageRegistration';
-import Timeline from './timeline/Timeline';
+import PageRegistration from "../Pages/PageRegistration";
+import PageTimeline from "../Pages/PageTimeline";
+import UserContext from "../contexts/UserContext";
+import Reset from "../styled-components/Reset";
 
 export default function App() {
-  return(
-    <BrowserRouter>
-      <Switch>
-        <Route path='/' exact>
-          <PageLogin/>
-        </Route>
-        <Route path='/sign-up' exact>
-          <PageRegistration/>
-        </Route>
-        <Route path='/timeline'exact>
-          <Timeline/>
-        </Route>
-      </Switch>
-    </BrowserRouter>
+  const [user, setUser] = useState('');
+  return (
+    <UserContext.Provider value={{user, setUser}}>
+      <Reset />
+      <BrowserRouter>
+        <Switch>
+          <Route path="/" exact>
+            <PageLogin />
+          </Route>
+          <Route path="/sign-up" exact>
+            <PageRegistration />
+          </Route>
+          <Route path="/timeline" exact>
+            <PageTimeline />
+          </Route>
+        </Switch>
+      </BrowserRouter>
+    </UserContext.Provider>
   );
 }
