+++ src/styled-components/Shared-Styled-Login-Registration.js
